// src/app/shared/model/update-master-cust/update-cust-emergency-obj.model.ts
export class UpdateCustEmergencyObj {
    CustEmergencyId: number;
    CustId: number;
    CustName: string;
    CustRelation: string;
    IdType: string;
    IdNo: string;
    IdExpiredDt: string;
    BirthPlace: string;
    BirthDate: string;
    Gender: string;
    Profession: string;
    Email: string;
    MobilePhn1: string;
    MobilePhn2: string;
    Address: string;
    Zipcode: string;
    AreaCode1: string;
    AreaCode2: string;
    AreaCode3: string;
    AreaCode4: string;
    City: string;
    RowVersion: string;
}